backend/pyproject.toml
[tool.poetry]
name = "trash-kan"
version = "0.1.0"
description = ""
authors = ["Lucas Franco <127855437+lucasfc-dev@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
requests = ">=2.32.5,<3.0.0"
oauthlib = ">=3.3.1,<4.0.0"
supabase = "^2.18.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
